Refactored containers
Discussed with @xyfi to move all SEMrush related redux into one container. Renamed SEMrushRelatedKeyphrases to RelatedKeyphrases to avoid confusion since it just returns the keyphrase
