+++ js/src/initializers/edit.js
@@ -30,7 +30,7 @@ import { link } from "../inline-links/edit-link";
 import PrePublish from "../containers/PrePublish";
 import DocumentSidebar from "../containers/DocumentSidebar";
 import PostPublish from "../containers/PostPublish";
-import SEMrushRelatedKeyphrases from "../containers/SEMrushRelatedKeyphrases";
+import SEMrushRelatedKeyphrases from "../containers/RelatedKeyphrases";
 
 const PLUGIN_NAMESPACE = "yoast-seo";
